move gitutils tests into dedicated test module; add test for git_location_to_raw
